+++ R/compat.R
@@ -35,7 +35,7 @@ compat <- function(r, target_version = package_version('3.0')) {
   r$dataTablesResults$dayOfTheWeek <- .fixDataFrameNames(r$dataTablesResults$dayOfTheWeek)
 
   # Backwards compatibility, typeConcepts and tableDateRange were merged into dateRangeByTypeConcept (v3)
-  if (!is.null(r$dataTablesResults$dateRangeByTypeConcept)) {
+  if (!is.null(r$dataTablesResults$dateRangeByTypeConcept$result)) {
     if (is.null(r$dataTablesResults$dateRangeByTypeConcept$result$FIRST_START_DATE)) {
       r$dataTablesResults$dateRangeByTypeConcept$result <- r$dataTablesResults$dateRangeByTypeConcept$result %>%
         mutate(
@@ -80,7 +80,7 @@ compat <- function(r, target_version = package_version('3.0')) {
   r$vocabularyResults$unmappedProcedures <- .fixDataFrameNames(r$vocabularyResults$unmappedProcedures)
   r$vocabularyResults$unmappedDevices <- .fixDataFrameNames(r$vocabularyResults$unmappedDevices)
   r$vocabularyResults$unmappedVisits <- .fixDataFrameNames(r$vocabularyResults$unmappedVisits)
-  if (!is.null(r$vocabularyResults$unmappedUnits)) {
+  if (!is.null(r$vocabularyResults$unmappedUnits$result)) {
     r$vocabularyResults$unmappedUnits <- .fixDataFrameNames(r$vocabularyResults$unmappedUnits)
     r$vocabularyResults$unmappedUnitsMeas$result <- r$vocabularyResults$unmappedUnits$result %>%
       filter(.data$DOMAIN == "measurement") %>%  # Renamed from TABLE to DOMAIN with fixDataFrameNames
@@ -97,6 +97,10 @@ compat <- function(r, target_version = package_version('3.0')) {
     r$vocabularyResults$unmappedUnitsObs <- .fixDataFrameNames(r$vocabularyResults$unmappedUnitsObs)
   }
 
+  if (!("unmappedEpisodes" %in% r$vocabularyResults)) {
+    r$vocabularyResults$unmappedEpisodes <- NULL
+  }
+
   r$vocabularyResults$mappedDrugs <- .fixDataFrameNames(r$vocabularyResults$mappedDrugs)
   r$vocabularyResults$mappedConditions <- .fixDataFrameNames(r$vocabularyResults$mappedConditions)
   r$vocabularyResults$mappedMeasurements <- .fixDataFrameNames(r$vocabularyResults$mappedMeasurements)
@@ -104,7 +108,7 @@ compat <- function(r, target_version = package_version('3.0')) {
   r$vocabularyResults$mappedProcedures <- .fixDataFrameNames(r$vocabularyResults$mappedProcedures)
   r$vocabularyResults$mappedDevices <- .fixDataFrameNames(r$vocabularyResults$mappedDevices)
   r$vocabularyResults$mappedVisits <- .fixDataFrameNames(r$vocabularyResults$mappedVisits)
-  if (!is.null(r$vocabularyResults$mappedUnits)) {
+  if (!is.null(r$vocabularyResults$mappedUnits$result)) {
     r$vocabularyResults$mappedUnits <- .fixDataFrameNames(r$vocabularyResults$mappedUnits)
 
     r$vocabularyResults$mappedUnitsMeas$result <- r$vocabularyResults$mappedUnits$result %>%
@@ -121,13 +125,16 @@ compat <- function(r, target_version = package_version('3.0')) {
     r$vocabularyResults$mappedUnitsMeas <- .fixDataFrameNames(r$vocabularyResults$mappedUnitsMeas)
     r$vocabularyResults$mappedUnitsObs <- .fixDataFrameNames(r$vocabularyResults$mappedUnitsObs)
   }
+
+  if (!("mappedEpisodes" %in% r$vocabularyResults)) {
+    r$vocabularyResults$mappedEpisodes <- NULL
+  }
+
   r <- .fixP_RECORDS(r)
 
-  # DED v1.0.5 has three additional columns
-  if (!('missing_quantity_exp_start_end_days_supply' %in% names(r$drugExposureDiagnostics$result))) {
-    r$drugExposureDiagnostics$result$missing_quantity_exp_start_end_days_supply <- NA
-    r$drugExposureDiagnostics$result$n_dose_and_missingness <- NA
-    r$drugExposureDiagnostics$result$median_daily_dose_q05_q95 <- NA
+  # DED v1.0.6 has three additional columns
+  if (!is.null(r$drugExposureDiagnostics$result)) {
+    r <- .fixDED(r)
   }
 
   # Performance
@@ -150,7 +157,11 @@ compat <- function(r, target_version = package_version('3.0')) {
 
 .get_cdmonboarding_version <- function(r) {
   if (is.null(r$cdmOnboardingVersion)) {
-    t <- as.data.frame(r$packinfo)
+    if (is.null(r$packinfo)) {
+      t <- r$performanceResults$packinfo
+    } else {
+      t <- as.data.frame(r$packinfo)
+    }
     version_string <- t[t$Package == 'CdmOnboarding', 'Version']
     return(package_version(version_string))
   }
@@ -220,5 +231,15 @@ compat <- function(r, target_version = package_version('3.0')) {
       r$vocabularyResults$drugMapping$result$P_RECORDS <- r$vocabularyResults$drugMapping$result$N_RECORDS / n_records * 100
     }
   }
+  return(r)
+}
+
+.fixDED <- function(r) {
+  if (!('missing_quantity_exp_start_end_days_supply' %in% names(r$drugExposureDiagnostics$result))) {
+    r$drugExposureDiagnostics$result$missing_quantity_exp_start_end_days_supply <- NA
+    r$drugExposureDiagnostics$result$n_dose_and_missingness <- NA
+    r$drugExposureDiagnostics$result$median_daily_dose_q05_q95 <- NA
+  }
+  r$drugExposureDiagnostics$packageVersion <- '1.0.5' # TODO: get from results$performanceResults$packinfo$Package
   return(r)
 }