+++ R/executeQuery.R
@@ -96,7 +96,7 @@ executeQuery <- function(
     ParallelLogger::logInfo(sprintf("> %s in %.2f secs", successMessage, duration))
   },
   error = function(e) {
-    ParallelLogger::logError(sprintf("%s", e))
+    ParallelLogger::logError(e)
     ParallelLogger::logError(sprintf("> Query failed. See '%s' for more details", errorReportFile))
   },
   finally = {