+++ extras/PackageMaintenance.R
@@ -9,11 +9,13 @@
 # 1. From develop, create a release-candidate branch from develop, e.g. release-v2.0
 # 1.1 Add release notes to NEWS.md
 # 2. Bump version in DESCRIPTION
+# 2.1 Run Document and Check. Check should finish without notes, warnings or errors.
+# 2.2 Commit and push changes to release-candidate branch
 # 3. Create Github PR to main
 # 4. Review and merge PR
 # 5. Create Github release on darwin-eu-dev, creating a tag as well.
 #    * make sure to target main
 # 6. Git pull
 # 7. Push to darwin-public git push --tags darwin-public main
-# 8. Merge main into developm, bump dev version in DESCRIPTION (add .900)
+# 8. Merge main into develop, bump dev version in DESCRIPTION (add .900)
 # 9. Create Github release on darwin-eu, reusing tag.